+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [master]
 ### Added
 
+- [api] Add support for the 'ttm' field in the query language. The field allows to filter on merged changes of a specific duration (in seconds).
 - [api] Expose the duration field of the change object. Duration is the time to merge delay in seconds.
 
 ### Changed
@@ -17,6 +18,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - [api] The self-merged count on the Activity page was incorrect and it is now properly calculated.
+- [crawler] Gerrit and Gitlab crawler compute negative duration for merged changes.
 
 ## [1.6.0] - 2022-06-20